requirements.txt
pandas>=2.0.0
numpy>=1.24.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
hdbcli>=2.0.0
requests>=2.28.0
openpyxl>=3.0.0
